pyproject.toml
[tool.poetry]
name = "DominionAI"
version = "0.1.0"
description = "Implements a Dominion buy strategy with genetic algorithms."
authors = ["Nate <minecraftcrusher100@gmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
pydominion = "^0.1.0"
numpy = "^1.23.0"

[tool.poetry.dev-dependencies]

[tool.pylint.messages_control]
disable = [
    "missing-class-docstring",
    "missing-module-docstring",
    "missing-function-docstring"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
